runtime: Document args[0] and the no-usable-args fallback
This follows Python's API, which has 'args' and 'executable' (our
PATH) for Popen [1]).  That allows most users (who don't need to
distinguish between args[0] and the executable path) to just use args,
while still providing a means to make that distinction (set 'path')
for folks who do need it.

This restores the ability to explicitly set args[0] independent of the
path, which was requested in #34 [2] and ack-ed by Micheal [3] and
Mrunal [4], but didn't match the examples that landed with #34.

[1]: https://docs.python.org/3/library/subprocess.html#subprocess.Popen
[2]: https://github.com/opencontainers/specs/pull/34#issuecomment-126116173
[3]: https://github.com/opencontainers/specs/pull/34#issuecomment-127335834
[4]: https://github.com/opencontainers/specs/pull/34#issuecomment-127350235

Signed-off-by: W. Trevor King <wking@tremily.us>