Move hooks from lifecycle.md to runtime.md
Hooks are only valid in runtime.json, so we should discuss them in the
runtime.json docs.  I'll add appropriate links from the lifecycle docs
in a bit, but linking into runtime.json is poor enough that I think it
deserves its own commit.  The only change here besides the copy/paste
is the addition of a blank line after the Markdown headers to match
the rest of the documentation.

Signed-off-by: W. Trevor King <wking@tremily.us>