lifecycle: Drop 'Runtime and Lifecycle' top heading
This page is just about the lifecycle now.  I'm not sure what the
original intention was, but when this header landed in runtime.md with
5d2eb180 (*: re-org the spec, 2015-06-24) there wasn't anything beyond
the lifecycle stuff in this file.  Now that it's lifecycle.md and
there are other files for our other content, I don't expect it to talk
about anything besides the lifecycle in the future.

Signed-off-by: W. Trevor King <wking@tremily.us>