state: Shift the state tree to it's own set of files
Michael was adding these to existing Go files to avoid having a lot of
files [1].  But 7232e4b (specs: introduce the concept of a
runtime.json, 2015-07-30, #88) split the runtime-configuration into
two file sets (one rooted in config.go for config.json and another
rooted in runtime_config.go for runtime.json).  In the face of a
two-set split for a single task (feeding the runtime while it
manipulates the container lifecycle), it seems odd to push the type
definition for a completely different task (sharing
container/application state with other tools) into an existing file
set.

[1]: https://github.com/opencontainers/specs/pull/87#issuecomment-128137602

Signed-off-by: W. Trevor King <wking@tremily.us>