Specify UTF-8 for all our JSON
I wish there was a cleaner reference for what UTF-8 was.  But [1]
seems too glib, and I can't find a more targetted link than just
dropping folks into a Unicode chapter (which is what [1] does):

  The Unicode Standard, Version 6.0, §3.9 D92, §3.10 D95 (2011)

With the current v8.0 (2015-06-17), it's still §3.9 D92 and §3.10 D95.

To get a place to say this for the runtime.json docs, I used config.md
as a template for the top-level header and blurb.  The "Host-specific
container configuration" phrasing comes from bundle.md, which has:

  The `runtime.json` file contains settings that are host specific...

While I was touching the config.md lead-in, I fixed 'container' ->
'bundle'.

[1]: https://en.wikipedia.org/wiki/UTF-8

Signed-off-by: W. Trevor King <wking@tremily.us>