change the user of the Process to a json.RawMessage
The type of `user` becomes `json.RawMessage` instead of platform aware
structure, so the Spec becomes a platform-independent structure.

It makes we can parse the config.json as a Spec before knowing
the os, and allows we move the linux related *.go to linux/.

The runtime should parse the user with the corresponding platform
specific User structure.

Signed-off-by: Lai Jiangshan <jiangshanlai@gmail.com>