.tools: Remove redundant r.Pass check when 'fail == 0'
The previous version of this script would always pass this check,
because when 'fail == 0' was true, all of the messages had passed.
That lead to logic like:

a. If all checks passed for the commit and the verbose flag is set,
   print all the messages (which all passed).
b. If all checks passed for the commit and the verbose flag is not set,
   don't print details about any messages (which all passed).
c. If any checks failed, print all the failed messages.

That wasn't printing successful checks when the verbose flag was true
but some checks failed.  This commit shifts the logic around to:

A. If the verbose flag is set, print messages for all checks, using
   TAP's "ok" and "not ok" to distinguish between per-check pass/fail
   [1].
B. If the verbose flag is not set, only print the failed messages (but
   still keep the "not ok" prefix).

In the non-verbose case, then (b/c) and (B) look the same (modulo my
"not ok" prefix adjustment).  And in the verbose case, I think (A) is
preferable to (a/c), because there's no reason to *not* print
successful checks when you've enabled the verbose flag.

[1]: http://testanything.org/

Signed-off-by: W. Trevor King <wking@tremily.us>