proto: initial protobuf for config
To have a less source based, and more consumable example of structures,
this is an initial pass at protobuf structures for the structures in config.go

There is some exercise needed in platform specific structures.

For the sake of example, the Makefile defaults to outputing golang
source, but has a 'c' target (`make c`) for C source as well.

Signed-off-by: Vincent Batts <vbatts@hashbangbash.com>