proto: initial protobuf for structures
To have a less source based, and more consumable example of structures,
this is an initial pass at protobuf structures for the structures in the
specs golang source.

There is some exercise needed in platform specific structures.

For the sake of example, the Makefile defaults to outputing golang
source, but has a 'c' target (`make c`) for C source, a `make py` and
`make all` target.

This User structure does not map to the cleanliness of the current go
structure, but will allow the definition to be all in one place,
regardless of the host that is doing the compilation.

All field rules to `optional` for now. Per vish and the docs,
https://developers.google.com/protocol-buffers/docs/proto?csw=1#specifying-field-rules
"! Required Is Forever"

There are a couple of concessions, but is pretty much lined up.

Signed-off-by: Vincent Batts <vbatts@hashbangbash.com>