proto: Platform type for User, and misc
This User structure does not map to the cleanliness of the current go
structure, but will allow the definition to be all in one place,
regardless of the host that is doing the compilation.

Switching field rules to `optional` for now. Per vish and the docs,
https://developers.google.com/protocol-buffers/docs/proto?csw=1#specifying-field-rules
"! Required Is Forever"

Also add a `make py` and `make all` target.

Signed-off-by: Vincent Batts <vbatts@hashbangbash.com>