.tools: Make -v output comply with TAP version 13
Using TAP [1] makes it easy to consume the output of this tool in a
variety of existing harnesses [2].

Only the verbose results are TAP compliant, mostly because folks
aren't likely to care about verbosity if they're just piping the
results into a TAP harness.  If we end up wanting the non-verbose
output to be TAP compliant, we'll have to switch to the trailing count
approach explained in the "Unknown amount and failures" section of the
spec [1], and we'll have to keep a running counter of failed checks
(instead of the current len(DefaultRules)-based approach).

[1]: http://testanything.org/tap-version-13-specification.html
[2]: http://testanything.org/consumers.html

Signed-off-by: W. Trevor King <wking@tremily.us>