.tools: Run 'go vet ...' on each commit
Instead of just checking the tip, run 'go vet ...' on each commit in
the series.  This makes sure we catch errors with earlier commits even
if they were fixed by subsequent commits.

The bulk of the complexity here is checking out a previous commit's
tree to a temporary directory, which is handled by the new
GitCheckoutTree.  I'd expect Git to be able to checkout an arbitrary
tree to an arbitrary directory, but strangely it doesn't seem to be
able to do that without side effects involving HEAD or the index.
Instead, I call 'git archive ...' and pipe the results into tar to do
the unpacking.  You *can* unpack an archive using native Go, but there
didn't seem to be a convenient tar.Reader.Extract(directory string)
method.  Shelling out to tar seemed more robust than writing a few
dozen lines of Go, so that's what I've done.  I'd expect shelling out
is less portable, but we really only need this to run on Travis'
Ubuntu environment [1].  We can work out a more portable approach
if/when we need to support this script on a system without a
compatible 'tar' command.

[1]: http://docs.travis-ci.com/user/ci-environment/#Virtualization-environments

Signed-off-by: W. Trevor King <wking@tremily.us>