.tools: Run 'golint ./...' on each commit
Instead of just checking the tip, run 'golint ./...' on each commit in
the series.  This makes sure we catch errors with earlier commits even
if they were fixed by subsequent commits.

I've added ValidateResult.Detail to hold stdout from ExecTree calls,
because golint just prints warnings to its stdout and always returns
an exit code of 0.  Then in the golint rule, I check vr.Detail and
fail any result with content there.  In the main block, I print YAML
blocks [1] with vr.Detail, so folks consuming the TAP can figure out
what golint was complaining about.  The YAML block gets printed when
the verbose flag is set or when there's an error (otherwise it's hard
to figure out what golint was complaining about in non-verbose mode).

The YAML library ensures we don't get bitten by YAML-sensitive
characters in the Detail string (e.g. colons).  The gopkg.in location
is a nifty site allowing us to link to a specific API of the project
(which is hosted on GitHub) [2].

[1]: http://testanything.org/tap-version-13-specification.html
[2]: http://godoc.org/gopkg.in/yaml.v2

Signed-off-by: W. Trevor King <wking@tremily.us>