use /etc/passwd in place of the User struct
The type of `user` becomes `string` instead of platform-specific
structure, so the Spec becomes a platform-agnostic structure.

It makes we can parse the config.json as a Spec before knowing the
os. It removes the User structure confliction for different platforms.

For linux-based container, parsing the /etc/passwd gives us a consistent
and more reliable way to assign the uid/gids to the process.

Signed-off-by: Lai Jiangshan <jiangshanlai@gmail.com>