runtime-config-linux: Require libseccomp constants
Don't punt readers to libseccomp, because this spec is supposed to be
a self-sufficient contract between the bundle-author and the runtime
[1].  The bundle-author shouldn't care if the runtime uses libseccomp,
or which version of libseccomp it uses.  But they can rely on the
listed constants being supported (and we may want to bring over brief
descriptions to go along with the constants themselves).

Runtime authors, on the other hand, may want to know that v2.2.3 of
libseccomp provides all the required constants.  I've added links to
the libseccomp 2.2.3 header from which these constants were extracted,
and that should be enough of a hint for bundle authors to use a
compatible libseccomp version (if they want to use libseccomp at all).

[1]: https://github.com/opencontainers/specs/pull/200/files#r40257217

Signed-off-by: W. Trevor King <wking@tremily.us>