runtime: Change 'destroys state' to 'removes container'
The runtime hasn't destroyed any of it's internal state before calling
the post-stop hooks (it still knows the state JSON it's going to pipe
to the hooks, and could still have the original configs in memory if
it wanted to keep that around).  What's changed by the time we have
the post-stop hooks (and the reason to keep an explicit container ID
around) is that the *kernel* state (namespaces and cgroups) might be
gone by the time the post-stop hooks are run (any PID namespace that
was created by the container process will certainly be gone by then).
This commit updates the wording to more clearly mean "after removing
namespaces and cgroups" and to avoid folks interpreting it as "after
releasing some internal-to-the-runtime-process memory".

Signed-off-by: W. Trevor King <wking@tremily.us>