Allow the namespace path to be container ID
In numerous cases, we need to share the container's namespace with
an existing container. In these cases, we have to resolve the namespace
file step by step. For example, the docker resolves it from the
information of the target container in several steps and saves it to the
config (now it is in-memory config, it will be runtime.json in future).
We can move the resolving operation to the oci implementations
and allow the namespace path to be container ID in the spec.

Reasons:
0) it really repsents "sharing the namespace with existing container".
   (it is the most important reason.)
1) the administrator writes the runtime.json much conveniently and
   less error. And it is better readability and the administrator
   is easier to check the runtime.json which is written by himself
   or by other human or by other tools.
2) we may have more orchestra tools than oci implementations,
   the resolving operation is better in implementations than
   in the tools.

Signed-off-by: Lai Jiangshan <jiangshanlai@gmail.com>