runtime-config-linux: Clarify resources vs. cgroupsPath
This should help clarify the cgroupsPath setting added in #137, which
was the subject of some confusion in opencontainers/runc#397.  Issues
I'm trying to clarify here:

* If you specify a cgroupsPath, is the container added to that path or
  a sub-cgroup underneath it [1]?  (This commit rules in favor of
  "added to that path")

* If you specify a cgroupsPath, can the runtime modify that cgroup
  [2]?  (This commit rules "yes, if 'resources' is specified",
  following [3] and the Go comment from #137 [4]).

To help make the distinctions clearer, I've added a facet list to help
folks think about the difference between cgroup creation, process
assignment, and resource configuration.  cgroupsPath is just about
cgroup creation and process assignment.  'resources' is just about
resource configuration.  I've listed out Mrunal's first three cases
[3] to be even clearer.  I stayed away from the "neither are set"
case, since I covered that fairly directly in #237, which that was
punted back to the list [5] and has seen no further interest.  So I'm
not clear on what the intended semantics are there, although Mrunal's
wording in [4] seems to agree with the proposal in #237.

[1]: https://github.com/opencontainers/runc/pull/397#issuecomment-156527638
[2]: https://github.com/opencontainers/runc/pull/397#issuecomment-157026547
[3]: https://github.com/opencontainers/runc/pull/397#issuecomment-157146714
[4]: https://github.com/opencontainers/specs/commit/429f936ff9665fd364a402af961389dc4ece6962#diff-34c30be66233f08b447fb608ea0e66bbR30
[5]: https://groups.google.com/a/opencontainers.org/d/msg/dev/qWHoKs8Fsrk/c9mv6qXtDAAJ
     Message-ID: <20151029194427.GA30073@odin.tremily.us>

Signed-off-by: W. Trevor King <wking@tremily.us>