runtime: Require 'start' to validate the bundle
This spec places RFC-2119 requirements on both bundles (bundle.md,
config.md, ...)  and runtimes (runtime.md, runtime-linux.md).  While
it's possible to envision a system where bundle validation is
decoupled and the runtime is free to blindly stumble through as far as
it can, it's likely that users will expect the runtime to perform
basic bundle validatation for them.  With the requirement landing in
this commit, the runtime MUST at least validate bundles against this
spec.  The runtime MAY also validate the bundle against the host
system's capabilities, but is not required to do so.

Signed-off-by: W. Trevor King <wking@tremily.us>