[Config] Mount options fstab format is for Linux
The fstab format (https://wiki.archlinux.org/index.php/Fstab) may be appropriate for Linux, but may not be appropriate for other OS platforms.  

This edit clarifies that mount options for Linux MUST be in fstab format but leaves the format of options as an open question for other operating systems to clarify.  

Signed-off-by: Rob Dolin <RobDolin@microsoft.com>