runtime: Replace "user-specified process" with "user-specified code" in 'create'
Because we *do* want to create the process at create-time, we just
want to idle it running runtime-specified code until start-time.  The
language I'm changing to already exists in the lifecycle and start
sections.

Since there were a number of links to the process spec, I replaced the
inline targets with a reference-style target.  And while I was doing
that, I added a missing close-paren to lifecycle step 2.

Signed-off-by: W. Trevor King <wking@tremily.us>