runtime: Support event buffering
This is all very generic, and I expect more details to land in the
runtime API specification.  Requirements around the buffer-request
semantics (e.g. "buffer for 15 seconds" or "buffer the last 5 events"
or "buffer the 'created' event") seemed out of place at the level of
detail in this specification.

The goal is to allow for:

  $ funC create --event-buffer created ID &
  $ funC event --event created ID && hook1 && hook2 && funC start ID
  $ fg

To support blocking on the event without racing on "maybe the
'created' event happened before the 'event' operation attached".

Given the small number of required events, this buffering should not
be a large resource concern, and '--event-buffer created' would only
ever require a single event to be buffered per container.

There is still a race on "maybe the container has already been
destroyed and a second container has been created with the same ID
before the 'event' operation attached", but that seems much less
likely (especially since the caller is free to pick UUIDs).

Signed-off-by: W. Trevor King <wking@tremily.us>