config-linux: Require no cgroup tweaks when linux.resources is unset
Or empty.  Using:

  "resources": {}

should mean the same thing as:

  "resources": null

or as not specifying 'resources' at all, so we can drop the "Do not
specify" requirement.

It's good to be clear about what leaving the properties unset/empty
means though.  I'd prefer a config-wide rule like [1], but if that is
too much to bite off I expect we can at least do that for cases where
a new container is joining an existing cgroup.

[1]: https://groups.google.com/a/opencontainers.org/forum/#!topic/dev/DWdystx5X3A
     Subject: Exposing platform defaults
     Date: Thu, 14 Jan 2016 15:36:26 -0800
     Message-ID: <20160114233625.GN6362@odin.tremily.us>

Signed-off-by: W. Trevor King <wking@tremily.us>