config: Convert process.rlimits from an array to an object
Rlimits do not need either ordering or repeat entries for a single
type.  Using an object leans on the new wording from eeaccfab
(glossary: Make objects explicitly unordered and forbid duplicate
names, 2016-09-27, #584) to make both of those points explicit.

Also add Solaris support.  I'm not entirely clear on this, because
while Solaris is POSIX-certified system and there is a Solaris man
page for setrlimit, Abhijeeth claims no Solaris support for rlimits
[1].

The additionalProperties object bit comes from [2,3], although it is
not documented in draft 4 of the JSON Schema RFC [4].

[1]: https://github.com/opencontainers/runtime-spec/pull/564#discussion_r79016149
[2]: https://spacetelescope.github.io/understanding-json-schema/reference/object.html#properties
[3]: https://tools.ietf.org/html/draft-wright-json-schema-validation-00#section-5.18
[4]: https://tools.ietf.org/html/draft-zyp-json-schema-04

Signed-off-by: W. Trevor King <wking@tremily.us>