asciidoc.conf: Automatically number unlabeled paragraphs
For labeled paragraphs, e.g.:

  [[foo]]
  Foo bar baz...

the paragraph ID (pid) will be "foo".  For unlabeled paragraphs, e.g.:

  Foo bar baz...

the paragraph label will be _para-#.

The implementation is based on the FAQ entry [1], but I've shuffled
things around to get the _para-# support and a linkable pilcrow (a la
Purple Pilcrows [3]).  It would still be nice to have logic to only
show pilcrows on hover (although I'm not sure how that would work in
PDF output).

[1]: http://asciidoc.org/faq.html#_how_do_i_number_all_paragraphs
[2]: http://www.tbray.org/ongoing/When/200x/2004/05/31/PurpleAgain

Signed-off-by: W. Trevor King <wking@tremily.us>