protocols: Pull protocols out of the README
The README will be a GitHub landing page, but these are part of the
spec proper.  I've quoted [1] the xrefs [2] for now, because the
targets haven't been transitioned over yet.

The only change I made to the content was re-ordering the Linux
portions to group the runtime references together (instead of having
runtime before the configs and runtime-linux after).

[1]: http://asciidoc.org/userguide.html#X81
     Monospaced text
[2]: http://asciidoc.org/userguide.html#_xref

Signed-off-by: W. Trevor King <wking@tremily.us>