use-cases: Pull use-cases out of the README
The README will be a GitHub landing page, but these are the
informative lead-in to the spec.  I've quoted [1] the xrefs [2] for
now, because the targets haven't been transitioned over yet.  In some
cases, I expect the appropriate caption to be filled in by the
(possibly autogenerated) xreflabel [3].

While the use-cases came after the definitions and protocols in the
old README, I've put them first in the AsciiDoc spec.  It reads better
to me with the informative use-cases first to orient readers before we
get down to business with the normative definitions and protocols.

[1]: http://asciidoc.org/userguide.html#X81
     Monospaced text
[2]: http://asciidoc.org/userguide.html#_xref
[3]: http://asciidoc.org/userguide.html#X30
     anchor

Signed-off-by: W. Trevor King <wking@tremily.us>