runtime: Transition from Markdown to AsciiDoc
There's a bit of hoop-jumping here for syntax highlighting.  asciidoc
(the command) can generate highted HTML (and the source-highlighter
line in oci-runtime-spec.asc sets it up to use Pygments).  However,
asciidoc's direct HTML output does not support xref default captions
[1].  asciidoc's DocBook output, on the other hand, *does* support
default captions [1].  Unfortunately, in this case asciidoc punts the
syntax highlighting to DocBook as well, passing through a
programlisting tag [2,3].  Unfortunately, syntax highlighting there
depends on an XSLTHL extension module which is only implemented in
Java [4,5], and that's not worth the trouble ;).

Highlighting is a bit easier for the PDF output, because while
asciidoc is still delegating to DocBook, DocBook is in turn delegating
to LaTeX and listings.  Listings v1.6 doesn't support JavaScript or
JSON, but it does support Python [6], and Python's dict, list, string,
and number syntax is very close to JSON's object, array, string, and
number syntax.  To setup that mapping, we copy the usual
asciidoc-dblatex.sty from the system, add a json -> python alias, and
tell dblatex to use our amended style.

Also bump the section numbering depth back up to DocBook's default 8
[7].  For some reason, the default a2x output seems to set this to 2.
And bump the number of ToC levels up from two to four [8].

An anchor's default captions [9] seem to be treated literally without
further parsing.  So [[created,`created`]] will lead to references
like a regularly-styled "`created`" instead of the intended monospace
"created".  Monospace isn't worth jumping through lots of hoops for
though, so I've just left the backticks off and leaned on the link
highlighting to set the should-be-literals off from surrounding text.

The [[create,create]] anchors avoid auto-generated captions like:

  Section 5.4.2, "Create"

because we usually use those references as stand-ins for an invokable
command (e.g. `funC create ...`).

[1]: http://asciidoc.org/userguide.html#_xref
[2]: http://asciidoc.org/source-highlight-filter.html#_docbook_outputs
[3]: http://docbook.org/tdg/en/html/programlisting.html
[4]: http://docbook.sourceforge.net/release/xsl/1.79.1/doc/html/highlight.source.html
[5]: https://sourceforge.net/p/xslthl/wiki/Home/
[6]: http://mirrors.ctan.org/macros/latex/contrib/listings/listings.pdf#page=13
[7]: http://docbook.sourceforge.net/release/xsl/1.79.1/doc/html/section.autolabel.max.depth.html
[8]: http://asciidoc.org/faq.html#_the_asciidoc_toclevels_attribute_does_not_work_with_docbook_outputs_why
[9]: http://asciidoc.org/userguide.html#X30
     anchor

Signed-off-by: W. Trevor King <wking@tremily.us>