State when the runtime should and must not apply Linux ambient capabilities
Ambient capabilities are a feature, since Linux 4.3, that enables capabilities
to be set on non root proesses directly. They are the only way to set these,
so are desirable in the case of "no new privileges" where suid binaries or
filesystem capabilities cannot be used as tis flag denies these operations,
and therefore there is no other way to apply capabilities to non root processes.

Without "no new privileges" users generally expect suid binaries or
filesystem capabilities to be the way to grant capabilities to non
root processes.

See https://github.com/opencontainers/runc/pull/1286 for the `runc` pull
and detailed explanation of the security issues in offering a choice here.

Signed-off-by: Justin Cormack <justin.cormack@docker.com>