config: Make default Linux filesystems an example
The MUST default-filesystem wording altered in 279c3c09 (linux: relax
filesystem requirements for container, 2017-01-23, #666) had read (to
me, anyway) as:

  The runtime MUST supply these even if the config doesn't call for
  them in mounts.

with 279c3c09 weaking it to:

  The runtime SHOULD supply these even if the config doesn't call for
  them in mounts.

But that's not very useful (callers that *need* a given mount will
still have to configure it explicitly).  However, one interpretation
of the 279c3c09 wording seems to be something like [1]:

  Config authors probably want to include mounts entries for these.

That's fine, and this commit tries to make that interpretation more
obvious by shifting the config recommendation over to the Linux
'mounts' example.

The values I'm using are straight from [2].

[1]: https://github.com/opencontainers/runtime-spec/pull/666#issuecomment-277067251
[2]: https://github.com/opencontainers/runtime-tools/pull/24

Signed-off-by: W. Trevor King <wking@tremily.us>