config: Clarify capabilities(7) as the canonical source of Linux caps
With the "valid values it chooses to not support" language from
718f9f3 (origin/pr/673) minor narrative cleanup regarding config
compatibility, 2017-01-30, #673), the runtime is clearly free to
support a subset of the platform's capabilities.  But the runtime
should not be free to change the semantics of valid values
(e.g. CAP_CHOWN should always mean the same thing on Linux, regardless
of which runtime you use).

Signed-off-by: W. Trevor King <wking@tremily.us>