README: Document the alternate-week meeting times
On Thu, Mar 09, 2017 at 08:37:55AM -0700, Chris Aniszczyk wrote [1]:
> summarizing the discussion, how about we just alternative the time like
> Jonathan discussed?
>
> I believe that's the only fair thing to do and it's reassuring to hear from
> people like Phil who would be able to make the time along with others.
>
> I'm fine with the current time (1000AEST/1500PST/0000CET) and then
> 0400AEST/0900PST/1800CET

Folks with a POSIX ‘date’ command can find the week number with [2]:

  $ date +%V

There may be some doubling up around the end of the year, but we're
usually canceling meetings around then anyway.

I'd be happier with meeting times anchored to UTC to make life easier
for folks outside of the US, but one change at a time ;).

[1]: https://groups.google.com/a/opencontainers.org/d/msg/dev/p0mTOspVgd0/mh7FYse2BAAJ
     Subject: Re: Moving the OCI Call (again)
     Date: Thu, 9 Mar 2017 08:37:55 -0700
     Message-ID: <CAJg1wMTCGEFRuKoKBEbUPdho82TVH8sPZdGORK_NA2vCNe+w9w@mail.gmail.com>
[2]: http://pubs.opengroup.org/onlinepubs/9699919799/utilities/date.html

Signed-off-by: W. Trevor King <wking@tremily.us>