runtime: Container-scope-wide uniqueness for container IDs
As written:

>  MUST be unique across all containers on this host.

seems unenforceable.  How will runC know what IDs are in use with
other runtimes on that host?  And how will a runtime in one namespace
know which IDs are used in another namespace?

The create-time language is much closer to being enforcable, by
invoking the container-scope idea.  To firm this up, I've adjusted the
wording to aligh with the name of the scope section, and I've linked
to that scope section.

With the create-time check in place, I don't think there's any purpose
to documenting the scope again in the state JSON section, so I've
dropped those two lines.

Signed-off-by: W. Trevor King <wking@tremily.us>