config-linux: RFC 2119 wording for disableOOMKiller
The previous wording tiptoes between background docs about kernel
behavior and runtime requirements, so it's not clear what the runtime
is required to do or how you'd compliance-test runtimes.  The new
wording uses MUST (NOT) wording for the various cases, so the runtime
responsibilities are clear.

The implementation-defined case avoids specifying when:

* The parent memory cgroup may have different values of
  oom_kill_disable, and
* The runtime may or may not create a new memory cgroup for the
  controller (because of the "MAY attach the container process to
  additional cgroup controllers" language).

Signed-off-by: W. Trevor King <wking@tremily.us>